fix auto-rename errors
