+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/+page.ts
@@ -1,4 +1,4 @@
-import { inlangSymbol } from "../../inlang"
+import { inlangSymbol } from "../../inlang.js"
 import type { PageLoad } from "./$types.js"
 
 export const load = (async ({ parent }) => {
