+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/about/+page.server.ts
@@ -1,4 +1,4 @@
-import type { PageServerLoad } from "./$types.js.js"
+import type { PageServerLoad } from "./$types.js"
 
 export const load = (async ({ locals }) => {
 	console.info("about/+page.server.ts", locals.i18n.i("welcome"))
