+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/inlang/+server.ts
@@ -1,5 +1,5 @@
 import { json } from "@sveltejs/kit"
 import type { RequestHandler } from "./$types.js"
-import { languages } from "../../inlang.server"
+import { languages } from "../../inlang.server.js"
 
 export const GET = (() => json(languages)) satisfies RequestHandler
