+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/inlang/[language]/+server.ts
@@ -1,7 +1,7 @@
 import { json } from "@sveltejs/kit"
 import type { RequestHandler } from "./$types.js"
-import { getResource } from "../../../inlang.server"
+import { getResource } from "../../../inlang.server.js"
 
-// eslint-disable-next-line unicorn/no-null
 export const GET = (({ params: { language } }) =>
+	// eslint-disable-next-line unicorn/no-null
 	json(getResource(language) || null)) satisfies RequestHandler
